requirements.txt
selenium==4.1.3
pytest==7.1.1
webdriver_manager
allure-pytest